PiperOrigin-RevId: 406444408

+++ jax/_src/scipy/optimize/bfgs.py
@@ -96,7 +96,7 @@ def minimize_bfgs(
 
   d = x0.shape[0]
 
-  initial_H = jnp.eye(d)
+  initial_H = jnp.eye(d, dtype=x0.dtype)
   f_0, g_0 = jax.value_and_grad(fun)(x0)
   state = _BFGSResults(
       converged=jnp.linalg.norm(g_0, ord=norm) < gtol,
